systemd: fix Wincompatible-pointer-type
The parameter is passed in safe_strtoint function.
The function expects its parameters in int32_t
data type. Therefore, c compiler throws a
-Wincompatible-pointer-type warning.

This patch changes data type of the parameter
to int32_t.

+++ compositor/systemd-notify.c
@@ -120,7 +120,7 @@ wet_module_init(struct weston_compositor *compositor,
 {
 	char *watchdog_time_env;
 	struct wl_event_loop *loop;
-	long watchdog_time_conv;
+	int32_t watchdog_time_conv;
 	struct systemd_notifier *notifier;
 
 	notifier = zalloc(sizeof *notifier);
